implement login register form & security for editing deleting post, add image in form for creating post

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,10 @@ use App\Post;
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show'] ]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -38,12 +42,26 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:19999'
         ]);
+
+        if($request->hasFile('cover_image')){
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            $filename = \pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $filenameToStore = $filename.'_'.time().'.'.$extension  ;
+            $path = $request->file('cover_image')->storeAs('public/cover_image', $filenameToStore);
+
+        } else {
+            $filenameToStore = 'no_image.jpg';
+        }
 
         $post = new Post();
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $filenameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Article crée');
@@ -71,6 +89,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts')->with('error', 'Vous n\'êtes pas autoriser à modifier cette page !!!');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -90,6 +113,7 @@ class PostsController extends Controller
 
         $post = Post::find($id);
         $post->title = $request->input('title');
+        
         $post->body = $request->input('body');
         $post->save();
 
@@ -105,6 +129,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts')->with('error', 'Vous n\'êtes pas autoriser à supprimer cette page !!!');
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success', 'Article supprimé');
